% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{end_task}
\alias{end_task}
\title{End an active task.}
\usage{
end_task(session, task_id)
}
\arguments{
\item{session}{A \code{Session} from \code{\link[=session]{session()}}.}

\item{task_id}{The task to end. If it's not the active one, the
bridge raises a non-fatal \code{NO_ACTIVE_TASK} warning.}
}
\value{
The \code{session}, invisibly.
}
\description{
End an active task.
}
