% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fernsicht-package.R
\docType{package}
\name{fernsicht-package}
\alias{fernsicht}
\alias{fernsicht-package}
\title{fernsicht: Watch Your R Code's Progress From Anywhere}
\description{
Wrap any long-running loop with \code{blick()} (or use the lower-level
\code{with_session()} / \code{\link[=session]{session()}} APIs) to stream live progress to a
shareable URL. Progress flows peer-to-peer via WebRTC; the package
wraps the \code{fernsicht-bridge} Go binary which is downloaded lazily
from GitHub releases on first use.
}
\details{
See \code{vignette("fernsicht")} for a walkthrough.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MuteJester/Fernsicht}
  \item \url{https://app.fernsicht.space/}
  \item Report bugs at \url{https://github.com/MuteJester/Fernsicht/issues}
}

}
\author{
\strong{Maintainer}: Thomas Konstantinovsky \email{thomaskon90@gmail.com} [copyright holder]

}
\keyword{internal}
